refactor(runtime): extract media tools into media module
The largest single-batch extraction yet — 12 functions and their two
constants/helpers move into tool_runner/media.rs:

- tool_media_describe / tool_media_transcribe (vision + audio
  understanding via media_understanding::MediaEngine)
- tool_image_generate (+ save_media_images_to_workspace,
  save_media_images_to_uploads helpers)
- tool_video_generate / tool_video_status (async video pipelines)
- tool_music_generate
- tool_text_to_speech (+ convert_to_ogg_opus, finish_tts_result helpers)
- tool_speech_to_text

SUPPORTED_AUDIO_EXTS_DOC is promoted to pub(super) because
builtin_tool_definitions in mod.rs reads it for the media_transcribe
and speech_to_text schema descriptions (so the agent-facing format
list stays in lockstep with audio_mime_from_ext). audio_mime_from_ext
itself is also pub(super) so the inline tests still reach it via
`use self::media::audio_mime_from_ext` from the tests module.

The internal helpers (save_media_images_*, convert_to_ogg_opus,
finish_tts_result) stay module-private because they have no out-of-
module caller.

mod.rs: 10,043 -> 9,219 LOC (-824). cargo check + cargo clippy --lib
both clean.

Refs #3710
